Users are now able to join a class and will populate the my classes tableview. isInstructor is being set on user(will need to clean up what a client can do through interface still).

+++ AnywhereFitness/Models/FitnessClassRepresentation.swift
@@ -12,7 +12,7 @@ struct FitnessClassRepresentation: Codable {
     var type:String
     var startTime:String
     var duration:String
-    var intensity: Double
+    var intensity: Int
     var location:String
-    var maxSize:Double
+    var maxSize: Int
 }
